LayoutTests/imported/w3c/web-platform-tests/mediacapture-record/MediaRecorder-detached-context-expected.txt


FAIL MediaRecorder creation with detached context assert_equals: expected "NotAllowedError" but got "NotSupportedError"
